One step forward and 2 steps back
Alright, most of the code seems to work now but it is removing only position 2 and nothing else..... My best guess is that the location things that I was hoping would work didn't and it is just reading from the beginning. Will look at it later but it seems to be looking good. Sorry for all the spam but I would say this is pretty much done minus the whole... replacing only two thing...I guess.